UpdateBudgetLineItem api integration successful

+++ app/src/main/java/com/decagonhq/decapay/feature/editbudgetlineitem/data/network/model/EditBudgetLineItemRequestBody.kt
@@ -1,5 +1,5 @@
 package com.decagonhq.decapay.feature.editbudgetlineitem.data.network.model
 
 data class EditBudgetLineItemRequestBody(
-    val amount: Int?
+    val amount: Double?
 )
